requirements.txt
transformers
faiss-cpu
sentence-transformers
streamlit
openai
torch
numpy